Fix Ensure Annotations in constructors
