+++ src/ensure_plugin/java/me/liliandev/ensure/transformerutils/TransformerUtil.java
@@ -1,6 +1,5 @@
 package me.liliandev.ensure.transformerutils;
 
-import com.sun.source.tree.AnnotationTree;
 import com.sun.source.tree.MethodTree;
 import com.sun.source.tree.VariableTree;
 import com.sun.tools.javac.code.TypeTag;
@@ -10,11 +9,8 @@ import com.sun.tools.javac.util.Context;
 import com.sun.tools.javac.util.List;
 import com.sun.tools.javac.util.Name;
 import com.sun.tools.javac.util.Names;
+import me.liliandev.ensure.EnsurePlugin;
 
-import java.lang.annotation.Annotation;
-import java.util.Map;
-import java.util.Optional;
-import java.util.ServiceLoader;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -22,7 +18,6 @@ import java.util.stream.Stream;
 public class TransformerUtil {
     private static final Set<String> PRIMITIVES = Stream.of(byte.class, short.class, char.class, int.class, long.class, float.class, double.class).map(Class::getName).collect(Collectors.toUnmodifiableSet());
 
-    private static final Map<String, Transformer> TRANSFORMERS = findTransformers();
 
     private static Context context = null;
 
@@ -52,7 +47,29 @@ public class TransformerUtil {
     public static void addCheck(MethodTree method, VariableTree argument, JCTree.JCExpression condition, String errorMessage) {
         JCTree.JCIf check = createCheck(argument, condition, getPos(argument, method), errorMessage);
         JCTree.JCBlock body = (JCTree.JCBlock) method.getBody();
-        body.stats = body.stats.prepend(check);
+        boolean shouldDelayCheck1Statement = false;
+        if (!body.stats.isEmpty()) {
+            JCTree.JCStatement firstStatement = body.stats.getFirst();
+            if (EnsurePlugin.className.contains("ConduitBlockEntity")) {
+                if (firstStatement instanceof JCTree.JCExpressionStatement potentialConstructor) {
+                    if (potentialConstructor.getExpression() instanceof JCTree.JCMethodInvocation methodInvocation) {
+                        if (methodInvocation.meth instanceof JCTree.JCIdent methodIdentifier && methodIdentifier.name.contentEquals("super")) {
+                            shouldDelayCheck1Statement = true;
+                        }
+                    }
+                }
+            }
+        }
+        if (shouldDelayCheck1Statement) {
+            List<JCTree.JCStatement> existingStatements = body.stats;
+            body.stats = List.of(existingStatements.getFirst());
+            body.stats = body.stats.append(check);
+            for (int i = 1; i < existingStatements.size(); i++) {
+                body.stats = body.stats.append(existingStatements.get(i));
+            }
+        } else {
+            body.stats = body.stats.prepend(check);
+        }
     }
     private static JCTree.JCIf createCheck(VariableTree argument, JCTree.JCExpression condition, int pos, String errorMessage) {
         Context context = TransformerUtil.getContext();
@@ -94,60 +111,11 @@ public class TransformerUtil {
 
 
 
-    public static void handle(VariableTree argument, MethodTree tree, String className) {
-        argument.getModifiers().getAnnotations()
-                .stream()
-                .map(TransformerUtil::matchTransformer)
-                .filter(tuple -> tuple.b() != null)
-                .forEachOrdered(tuple -> tuple.b().transform(argument, tree, tuple.a(), className));
-    }
-    public static void handle(MethodTree tree, String className) {
-        tree.getModifiers().getAnnotations()
-                .stream()
-                .map(TransformerUtil::matchTransformer)
-                .filter(tuple -> tuple.b() != null)
-                .forEachOrdered(tuple -> tuple.b().transform(null, tree, tuple.a(), className));
-    }
-    public static String getName(AnnotationTree annotation) {
 
-        if (annotation.getAnnotationType() instanceof JCTree.JCIdent ident) {
-            if (ident.sym == null) {
-                //some java.lang annotations don't get a symbol as they are in java.lang
-                return null;
-            }
-            return ident.sym.flatName().toString();
-        }
-        return null;
-    }
 
     public static Name getParameterId(Names symbolsTable, VariableTree parameter) {
         String parameterName = parameter.getName().toString();
         return symbolsTable.fromString(parameterName);
     }
 
-    private static Tuple<AnnotationTree, Transformer> matchTransformer(AnnotationTree annotation) {
-        return new Tuple<>(annotation, TRANSFORMERS.get(getName(annotation)));
-    }
-
-
-    public static Map<String, Transformer> findTransformers() {
-        return ServiceLoader.load(Transformer.class, TransformerUtil.class.getClassLoader())
-                .stream()
-                .map(transformer -> findMarker(transformer.type()).map(annotation -> new Tuple<>(annotation, transformer.get())))
-                .filter(Optional::isPresent)
-                .map(Optional::get)
-                .collect(Collectors.toMap(tuple -> tuple.a.getName(), tuple -> tuple.b));
-    }
-
-    private static  Optional<Class<? extends Annotation>> findMarker(Class<? extends Transformer> clazz) {
-        Handler annotation = clazz.getAnnotation(Handler.class);
-        if (annotation == null) {
-            return Optional.empty();
-        }
-        return Optional.of(annotation.annotation());
-    }
-
-    private record Tuple<A, B>(A a, B b) {
-
-    }
 }
